PMM-7574 move EmptyBlock to shared
