+++ public/app/percona/shared/EmptyBlock/EmptyBlock.styles.ts
