+++ public/app/percona/shared/EmptyBlock/EmptyBlock.types.ts
